refactor(snmp): split neighbor_mib.go (1257 → 244 lines) by MIB
neighbor_mib.go was 1257 lines mixing the initialization of six MIBs
in one file. Same-package split by MIB (each MIB is its own concern):

  neighbor_mib.go (slim) constants + initializeNeighborMIBs orchestrator
  mib_if.go              IF-MIB / IF-MIB-X tables + counters
  mib_ip.go              IP-MIB + ipNetToMediaTable (ARP)
  mib_bridge.go          BRIDGE-MIB (dot1d Base/STP/Tp) + macToOIDIndex
  mib_discovery.go       LLDP local + remote + CDP cache entries
  mib_helpers.go         parseMACBytes / getInterfaceSpeed / getCapabilitiesBitfield

go test ./internal/protocols/snmp/... passes; golangci-lint clean.
